Ensure all kernel references are in highmem, enable exec syscall
Up until now, multiple references to kernel data retained pointers in lowmem.
The root cause is that llvm makes $ebx point to the PLT at all times, and it
is initialized from the starting instruction pointer. Since $eip starts low
with paging disabled, the PTL also gets located below 0xc0000000. This
cascades to nearly every pointer reference stored somewhere in memory.

With the introduction of the exec syscall, user processes unmap the kernel from
the first 4MiB of memory before loading a new one. This broke a bunch of
things, with the most complicated being the vtable lookups for dynamic Traits.
To fix this, $ebp is moved beyond 0xc0000000 once paging is enabled. This is
a hacky fix that will probably persist until the bootloader gets modified to
enable paging and start the entire kernel in highmem.

To demonstrate exec with a simple binary file, a program that prints to COM1
is included in the initfs.