Add dup and dup2 syscalls
Untested, but this changes the way file handles are reused and allocated when files are opened. The dup2 command allows specifying the handle to use, so this separates out allocation from actually setting something at that index. Old handles are now reused if they've been closed. When all handles referencing the same open file are closed, the fs is told it can close that instance of the file.