Manage process state, and introduce simple IPC handling
To rectify the complicated blocking states that were appearing on the old
branch, the process state has been reimagined.
Beyond signal handling, or explicitly waiting on a child to exit, all blocking
will occur as a result of IPC reads. For example, communicating with a disk
driver could be modeled as sending a read request (what and how much to read,
and where to place the results), and then waiting on an IPC response that tells
the initial process how much was read, or what error occurred.

IPC is implemented as a per-process inbox. Messages that are sent to the
process wait in this queue until the process issues an IPC read syscall.
Reading returns the oldest message in the queue, if one exists, and also
tells the process if there are any other messages waiting to be read.

In order to prevent the queue from growing too large, especially in processes
that don't intend to read their queues, entries in the queue have an expiration
time. Each time the queue is modified, the entries at the front of the queue
have their expiration time inspected, and they are removed if it has passed.