Adding stacks to processes, enabling new fork and switch behavior
This dramatically rethinks the way task switching happens at the kernel level.
Previously, stacks existed at the same location and relied on page mapping to
change what it pointed to. The guarantees this created were marginal, and made
modification of other processes' stacks much harder.
There's more than enough virtual memory to support large kernel stacks for 32k
processes at the same time. With this in mind, a chunk of memory has been
reserved at the top of kernel space for process stacks. When a new process is
created, a 4-page stack gets reserved. The frames for those pages are mapped
in kernel space for all processes. The lowest page remains unmapped to act as
a guard page and prevent one process from clobbering another's stack.

The organization of processes has changed as well. Previously, each process
contained many locked fields, but this led to the potential for deadlocks.
Now, the map of existing processes places the entire struct in a lock,
guaranteeing that if one field is accessible, all the fields are accessible.
It also makes it easier to write methods that modify the internals of the
struct.

Forking is significantly simplified as well. Previously the stack was copied
from the parent to the child, and the child unwound the stack to return to
the fork() call entrypoint. It turns out this is unnecessary -- the child only
needs to put the userspace entrypoint on its stack and perform an `iret` call.