Re-enable jumping to userspace on the new Task implementation
Demonstrates successfully jumping to usermode and running code there. Since
the TSS hasn't been set up yet, exceptions and syscalls won't work.

Moving forward, this hook won't actually be necessary. There will only be two
ways that the kernel moves to userspace. The first involves the init process:
the kernel will create a new process, perform an `exec()` call on it to load
the relevant memory segments, and then push the entrypoint onto its kernel
stack. Marking that new process as "current" and calling `iretd` will jump
into the init userspace program.
The second way to jump to userspace will occur anytime a syscall completes.
Most of these will simply rewind the stack and `iretd` back to the call site.
For processes that were created via a fork, a special return-to-callsite hook
will be created and pushed onto the stack.