Actually use memory isolation for the new process architecture
Now processes will have their own dedicated page directories. Each directory
contains a self-mapping at index 1023, and all other entries above 0x2ff (the
last userspace entry) are shared between all directories. This allows all
proceses to share the same global kernel state, as well as access to the stacks
of all other processes.

Manipulation of a page that isn't in the current memory mapping has been
simplified. The top "stack" is actually a reserved space for scratch pages that
are unmapped in the current process, and the new `UnmappedPage` struct makes it
easy to allocate and deallocate those scratch pages without needing to know
where they are, or how big the scratch space is.

As execution of userspace programs is re-implemented, this will allow us to
trigger page faults, and eventually page in program code and data on demand.
The next few diffs will update the page fault logic to use the new process
structures.