Allow paging executable files on demand
Begin rewriting the internal of the page fault handler. When a page fault
occurs in userspace because of an unmapped page, it queries the current
process to determine if that memory region is part of the executable, the
brk heap, or a mmap-ed page. If it's part of the heap, a zero-filled frame
will be mapped to the accessed page. If it's part of the executable, all
sections overlapping with that page will be copied to a new frame.

If a mapping was produced, it returns to the callsite and execution should
continue as normal. If not, it falls through to generate a segfault for the
current process.