Run an installed interrupt handler when the trigger is raised
This will enter usermode for that process and jump to the handler address.
Upon returning from the handler, it will attempt to jump to a "magic number"
address and let the page fault handler know that we are returning from an
interrupt.
At this point we have no way to handle that return, so we just spin forever.
This can be resolved in follow-up commits that will stash all process state
for every interrupt handler.