Successfully enter an installed interrupt handler, and return upon completion
When the interrupt occurs, all state is stored in the current process and the
kernel jumps to the interrupt handler, if installed. When that function
returns, the kernel picks up the magic number and recognizes that the int is
complete. It uses the stored info to jump back to the original process without
any hiccups.