Reorganizing the project around a clearer roadmap
As more systems get implemented and re-implemented, the OS is becoming just
another unix-like with no significantly unique traits. Also, DOS features have
been pushed to the side as more modern features were developed. I'm rethinking
the purpose of the project, and settling on a more explicit vision.

POSIX compatibility is now an explicit non-goal. Instead, the core goal is
hosting of DOS programs in a multitasking, 32-bit protected mode environment.
The features that will be developed first are focused on creating this host
environment, and making it compatible with as many legacy programs as possible.

Carried over from the old system is the idea that drivers should live in user
mode. Concepts like IPC message-passing and the IPIOA will continue to get
built out. However, implementation of filesystem and device drivers will be
secondary to fully covering the DOS API surface area.