Implement blocking reads for devices
Device drivers have been modified to use explicit IO handles.
There is now a reusable trait for implementing blocking IO, based on how it
was done with the COM driver.
To demonstrate blocking reads for the TTY driver, a COMMAND shell has begun
to take shape. It reads from the TTY until a newline is reached or the buffer
is full, and then prints another prompt.