Add concept of parent vterm, use it for gfx mode setting
Processes spawned from a vterm (as opposed to from the init process) will now
have a vterm associated with them. Anything they do to modify the environment
will be mapped to that specific vterm.
The graphics mode-setting syscall has been updated to reflect this. If a
process is not currently in the active vterm, and requests a video mode
change, those changes will only take place when its vterm is made active.

As future diffs add support for DOS modes, this will allow the task execution
code to properly modify the right vterm on program start and end.