Enable editing video memory from DOS programs
When a DOS program is loaded, its VTerm memory is mapped to the appropriate
places in low-mem. When a DOS program attempts to access video memory that
hasn't been accessed yet, a new backup frame will be allocated within the
vterm.
If a DOS program attempts to change video mode via INT 0x10, it will be
intercepted and handled appropriately.